Fix delete data & templating

+++ code.php
@@ -3,18 +3,21 @@ session_start();
 require 'dbcon.php';
 
 // Delete function
-$student_id = mysqli_real_escape_string($con, $_POST['delete_student']);
-$query = "DELETE FROM students WHERE id='$student_id'";
-$query_run = mysqli_query($con, $query);
-
-if($query_run){
-    $_SESSION['message'] =  'Student Data Deleted Successfully';
-    header('Location: index.php');
-    exit(0);
-}else{
-    $_SESSION['message'] =  'Student Data Not Deleted';
-    header('Location: index.php');
-    exit(0);
+if(isset($_POST['delete_student'])){
+    $student_id = mysqli_real_escape_string($con, $_POST['delete_student']);
+    $query = "DELETE FROM students WHERE id='$student_id'";
+    $query_run = mysqli_query($con, $query);
+    
+    if($query_run){
+        $_SESSION['message'] =  'Student Data Deleted Successfully';
+        header('Location: index.php');
+        exit(0);
+    }else{
+        $_SESSION['message'] =  'Student Data Not Deleted';
+        header('Location: index.php');
+        exit(0);
+    }
+    
 }
 
 // Update function
